Polish mobile composer controls
