+++ packages/app/src/components/composer.tsx
@@ -37,6 +37,7 @@ import {
   DraftAgentStatusBar,
   type DraftAgentStatusBarProps,
 } from "./agent-status-bar";
+import { ComposerCompactPreferencesExtras } from "./composer-compact-preferences-extras";
 import { ContextWindowMeter } from "./context-window-meter";
 import { useImageAttachmentPicker } from "@/hooks/use-image-attachment-picker";
 import { useSessionStore } from "@/stores/session-store";
@@ -800,82 +801,6 @@ function ComposerRightControlsSlot({
       {compactOverflowControls}
       {shouldShowVoiceModeButton ? <ComposerVoiceModeButton {...voiceProps} /> : null}
       {cancelButton}
-    </View>
-  );
-}
-
-function formatCompactTokenCount(value: number): string {
-  if (value >= 1_000_000) return `${Math.round(value / 1_000_000)}m`;
-  if (value >= 1_000) return `${Math.round(value / 1_000)}k`;
-  return Math.round(value).toString();
-}
-
-function formatContextWindowDetail(maxTokens: number | null, usedTokens: number | null): string {
-  if (maxTokens === null || usedTokens === null || maxTokens <= 0) {
-    return "Not available";
-  }
-  const percentage = Math.round((usedTokens / maxTokens) * 100);
-  return `${percentage}% used, ${formatCompactTokenCount(usedTokens)} / ${formatCompactTokenCount(maxTokens)} tokens`;
-}
-
-interface ComposerCompactPreferencesExtrasProps {
-  contextWindowMaxTokens: number | null;
-  contextWindowUsedTokens: number | null;
-  isVoiceModeForAgent: boolean;
-  hasAgent: boolean;
-  isConnected: boolean;
-  isVoiceSwitching: boolean;
-  handleToggleRealtimeVoice: () => void;
-}
-
-function ComposerCompactPreferencesExtras({
-  contextWindowMaxTokens,
-  contextWindowUsedTokens,
-  isVoiceModeForAgent,
-  hasAgent,
-  isConnected,
-  isVoiceSwitching,
-  handleToggleRealtimeVoice,
-}: ComposerCompactPreferencesExtrasProps): ReactElement {
-  const voiceDisabled = !hasAgent || !isConnected || isVoiceSwitching || isVoiceModeForAgent;
-  const voiceRowStyle = useCallback(
-    ({ pressed }: PressableStateCallbackType) => [
-      styles.compactSheetRow,
-      pressed && styles.compactSheetRowPressed,
-      voiceDisabled && styles.compactSheetRowDisabled,
-    ],
-    [voiceDisabled],
-  );
-  const contextMeter =
-    contextWindowMaxTokens !== null && contextWindowUsedTokens !== null ? (
-      <ContextWindowMeter maxTokens={contextWindowMaxTokens} usedTokens={contextWindowUsedTokens} />
-    ) : (
-      <View style={styles.compactSheetIconSlot} />
-    );
-  return (
-    <View style={styles.compactSheetExtraSection}>
-      <View style={styles.compactSheetRow}>
-        {contextMeter}
-        <View style={styles.compactSheetTextGroup}>
-          <Text style={styles.compactSheetTitle}>Context window</Text>
-          <Text style={styles.compactSheetDetail}>
-            {formatContextWindowDetail(contextWindowMaxTokens, contextWindowUsedTokens)}
-          </Text>
-        </View>
-      </View>
-      <Pressable
-        disabled={voiceDisabled}
-        onPress={handleToggleRealtimeVoice}
-        style={voiceRowStyle}
-        accessibilityRole="button"
-        accessibilityLabel={isVoiceModeForAgent ? "Voice mode active" : "Enable Voice mode"}
-      >
-        <ThemedAudioLines size={ICON_SIZE.md} uniProps={iconForegroundMutedMapping} />
-        <View style={styles.compactSheetTextGroup}>
-          <Text style={styles.compactSheetTitle}>Voice mode</Text>
-          <Text style={styles.compactSheetDetail}>{isVoiceModeForAgent ? "On" : "Off"}</Text>
-        </View>
-      </Pressable>
     </View>
   );
 }
@@ -1839,45 +1764,6 @@ const styles = StyleSheet.create((theme: Theme) => ({
   realtimeVoiceButtonActive: {
     backgroundColor: theme.colors.palette.green[600],
     borderColor: theme.colors.palette.green[800],
-  },
-  compactSheetRow: {
-    flexDirection: "row",
-    alignItems: "center",
-    gap: theme.spacing[3],
-    paddingHorizontal: theme.spacing[4],
-    paddingVertical: theme.spacing[3],
-    borderRadius: theme.borderRadius.lg,
-    borderWidth: 1,
-    borderColor: theme.colors.surface2,
-    backgroundColor: theme.colors.surface0,
-  },
-  compactSheetExtraSection: {
-    gap: theme.spacing[2],
-  },
-  compactSheetRowPressed: {
-    backgroundColor: theme.colors.surface2,
-  },
-  compactSheetRowDisabled: {
-    opacity: 0.5,
-  },
-  compactSheetIconSlot: {
-    width: 28,
-    height: 28,
-  },
-  compactSheetTextGroup: {
-    flex: 1,
-    minWidth: 0,
-    gap: theme.spacing[1],
-  },
-  compactSheetTitle: {
-    color: theme.colors.foreground,
-    fontSize: theme.fontSize.base,
-    fontWeight: theme.fontWeight.semibold,
-  },
-  compactSheetDetail: {
-    color: theme.colors.foregroundMuted,
-    fontSize: theme.fontSize.sm,
-    fontWeight: theme.fontWeight.normal,
   },
   iconButtonHovered: {
     backgroundColor: theme.colors.surface2,
